+++ src/PhotoMigrator.py
@@ -128,8 +128,6 @@ def main():
 def pre_parse_args():
     # Verificar si el script se ejecutó con un solo argumento que sea una ruta de una carpeta existente
     if len(sys.argv) >= 2 and os.path.isdir(sys.argv[1]):
-        # print(f"{GV.TAG_INFO}Valid folder detected as input: '{sys.argv[1]}'")
-        # print(f"{GV.TAG_INFO}Executing Google Takeout Photos Processor Feature with the provided input folder...")
         custom_print(f"Valid folder detected as input: '{sys.argv[1]}'", log_level=logging.INFO)
         custom_print(f"Executing Google Takeout Photos Processor Feature with the provided input folder...", log_level=logging.INFO)
         sys.argv.insert(1, "--google-takeout")
@@ -140,7 +138,6 @@ def pre_parse_args():
             root = tk.Tk()
             root.withdraw()
             return filedialog.askdirectory(title="Select the Google Takeout folder to process")
-
         try:
             import tkinter as tk
             from tkinter import filedialog
@@ -148,33 +145,26 @@ def pre_parse_args():
         except ImportError:
             TKINTER_AVAILABLE = False
 
-        # print(f"{GV.TAG_INFO}No input folder provided. By default, the Google Takeout Photos Processor feature will be executed.")
         custom_print(f"No input folder provided. By default, the Google Takeout Photos Processor feature will be executed.", log_level=logging.INFO)
         has_display = os.environ.get("DISPLAY") is not None or sys.platform == "win32"
-        selected_folder = None
 
         if has_display and TKINTER_AVAILABLE:
-            # print(f"{GV.TAG_INFO}GUI environment detected. Opening folder selection dialog...")
             custom_print(f"GUI environment detected. Opening folder selection dialog...", log_level=logging.INFO)
+            custom_print(f"Please, Select the Google Takeout folder to process...", log_level=logging.INFO)
             selected_folder = select_folder_gui()
         else:
             if not TKINTER_AVAILABLE and has_display:
-                # print(f"{GV.TAG_WARNING}Tkinter is not installed. Falling back to console input.")
                 custom_print(f"Tkinter is not installed. Falling back to console input.", log_level=logging.WARNING)
             else:
-                # print(f"{GV.TAG_INFO}No GUI detected. Using console input.")
                 custom_print(f"No GUI detected. Using console input.", log_level=logging.INFO)
-            # print(f"Please type the full path to the Takeout folder:")
-            custom_print(f"Please type the full path to the Takeout folder:", log_level=logging.WARNING)
+            custom_print(f"Please type the full path to the Takeout folder to process:", log_level=logging.WARNING)
             selected_folder = input("Folder path: ").strip()
 
         if selected_folder and os.path.isdir(selected_folder):
-            # print(f"{GV.TAG_INFO}Folder selected: '{selected_folder}'")
             custom_print(f"Folder selected: '{selected_folder}'", log_level=logging.INFO)
             sys.argv.append("--google-takeout")
             sys.argv.append(selected_folder)
         else:
-            # print(f"{GV.TAG_ERROR}No valid folder selected. Exiting.")
             custom_print(f"No valid folder selected. Exiting.", log_level=logging.ERROR)
             sys.exit(1)